+++ structure/goods/Good.py
