+++ structure/places/Place.py
